requirements.txt
aiogram>=3.17.0,<4.0.0
httpx>=0.27.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
